requirements.txt
django==3.2.24
djangorestframework==3.12.4
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
django-versatileimagefield==2.0.0
Pillow>=8.0.0,<10.0.0
python-magic-bin==0.4.14; platform_system == "Windows"

# django-versatileimagefield
# python-magic-bin==0.4.14
# Django==3.2.24
# pillow>=8.0.0
# djangorestframework==3.12.4
# django-cors-headers==4.4.0
# djangorestframework-simplejwt==5.3.1
# drf-yasg==1.21.8